requirements.txt
telebot==0.0.5
flask==2.3.3
pillow==10.0.1
numpy==1.24.3
gunicorn==21.2.0